pyproject.toml
[project]
name = "skill-thief"
version = "0.1.0"
description = "Simple CLI to install and update Agent Skills from git or local paths"
authors = [{name = "Skill Thief Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "typer>=0.12",
  "PyYAML>=6.0",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = ["behave>=1.2.6"]

[project.scripts]
skill-thief = "skill_thief.cli:app"

[project.urls]
Homepage = "https://github.com/ryan-porter/the-skill-thief"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.behave]
definitions = ["features/steps"]
